Changes to the plugin interface
With the previous version of the plugin you created an API client object
by passing provider_key AND the corresponding application keys.
Now you only pass the provider_key and the application keys are passed
when the method is invokated (authrep, authorize or report). That's how
the Ruby or Java plugins work.

It is backwards compatible with the older version. You can check it by
running the old test suite (tests.py) over the new plugin version.

Other changes:
- new suite of tests to cover the new way of using the plugin
  (tests-v2.py)
- added test coverage for the user_key auth pattern
- new examples and extended documentation